+++ kernel/kernel-api/src/main/java/io/delta/kernel/internal/table/ResolvedTableBuilderInternalImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.delta.kernel.internal;
+package io.delta.kernel.internal.table;
 
 import static io.delta.kernel.internal.DeltaErrors.wrapEngineExceptionThrowsIO;
 import static io.delta.kernel.internal.util.Preconditions.checkArgument;
